Implemented user story:     A developer can run 'check' without installing dependencies in their global library.
The approach is to add a new "bundle" method which installs package dependencies into a ".Rbundle" library under a user's home directory (see bundle_library). The CHECK command is then run against that library and not the user's global library.

Also in this commit:
* Bump up devtools version to 0.7.2
* Extract the 'R' function into it's own file and export it.
* Extract install_deps into it's own file and export it.
* Create a new package setting called "Repositories", which explicitly sets the allowed repositories as opposed to using global user settings.
* Document and export the system_check function.
* Test new 'bundle' function with two mock packages: simple-dependencies, no-dependencies.

The design of this feature aims to mimic what Bundler does for Ruby and maven does for Java. For some background for why this stuff might make sense to have in R, see http://stackoverflow.com/questions/7730180/dependency-management-in-r